backend/requirements.txt
# ============================================================================
# Core FastAPI Dependencies
# ============================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# ============================================================================
# Database (PostgreSQL with AsyncIO)
# ============================================================================
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
psycopg2-binary==2.9.9

# ============================================================================
# Authentication & Security
# ============================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
pyjwt==2.8.0

# ============================================================================
# File Handling & Validation
# ============================================================================
python-magic==0.4.27
pillow==10.1.0
aiofiles==23.2.1

# ============================================================================
# AI/ML Services
# ============================================================================
# Google Gemini AI
google-generativeai==0.3.1

# OpenAI (fallback)
openai==1.3.7

# HuggingFace Transformers (for sentiment analysis)
transformers==4.35.2
torch==2.1.1
sentencepiece==0.1.99

# Speech-to-Text (Whisper)
openai-whisper==20231117

# ============================================================================
# Redis (Caching & Sessions)
# ============================================================================
redis==5.0.1
hiredis==2.2.3

# ============================================================================
# HTTP Client & Web Scraping
# ============================================================================
httpx==0.25.2
aiohttp==3.9.1

# ============================================================================
# Background Tasks (Optional - Celery)
# ============================================================================
celery==5.3.4
flower==2.0.1

# ============================================================================
# Email Services (Optional)
# ============================================================================
fastapi-mail==1.4.1

# ============================================================================
# Monitoring & Logging
# ============================================================================
python-json-logger==2.0.7

# ============================================================================
# Testing
# ============================================================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# ============================================================================
# Development Tools
# ============================================================================
python-dotenv==1.0.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

# ============================================================================
# PDF & Document Processing (for Resume Analysis)
# ============================================================================
pypdf2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# ============================================================================
# Data Validation & Serialization
# ============================================================================
email-validator==2.1.0

# ============================================================================
# CORS & Security
# ============================================================================
python-cors==1.0.0

# ============================================================================
# Rate Limiting (Optional)
# ============================================================================
slowapi==0.1.9

# ============================================================================
# WebSocket Support (for real-time features)
# ============================================================================
python-socketio==5.10.0
websockets==12.0

# ============================================================================
# Production Server (Alternative to Uvicorn)
# ============================================================================
gunicorn==21.2.0
